bot/requirements.txt
aiohttp~=3.8.1
aiogram~=2.20
python-dotenv~=0.19.2
pycups~=2.0.1
pycups-notify~=0.0.5
prometheus-client~=0.11.0
unoserver~=1.2
PyPDF4~=1.27.0